+++ backend/commons/src/main/java/com/engineeringthesis/commons/dto/visit/ResultsReportDTO.java
@@ -17,7 +17,7 @@ public class ResultsReportDTO {
     @Size(max = 300, message = "Result description must contain maximum of 300 characters!")
     private String resultDescription;
 
-    private Integer numericalResult;
+    private Float numericalResult;
 
     @Size(max = 50, message = "Unit must contain maximum of 50 characters!")
     private String unit;
